# backend/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GestorPlus API",
        "description": "Documentación Swagger para la API de GestorPlus",
        "contact": {
            "name": "Soporte GestorPlus",
            "email": "soporte@tudominio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Servidor local"
        },
        {
            "url": "https://api.tudominio.com/",
            "description": "Servidor producción"
        }
    ],
    "paths": {
...
        "/permisos/solicitar": {
            "post": {
                "tags": [
                    "Permiso"
                ],
                "summary": "Solicitar permiso",
                "description": "Permite al usuario autenticado solicitar un permiso.",
                "operationId": "8ce6747dd44fa3eb259cf1251ff4a094",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motivo",
                                    "fechaInicio",
                                    "fechaFin"
                                ],
                                "properties": {
                                    "motivo": {
                                        "type": "string",
                                        "example": "Cita médica"
                                    },
                                    "fechaInicio": {
                                        "type": "string",
                                        "example": "2024-06-01"
                                    },
                                    "fechaFin": {
                                        "type": "string",
                                        "example": "2024-06-02"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso solicitado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Permiso solicitado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error al solicitar el permiso"
                    }
                }
            }
        },
        "/permisos/aceptar": {
            "patch": {
                "tags": [
                    "Permiso"
                ],
                "summary": "Aceptar permisos",
                "description": "Acepta uno o varios permisos por sus IDs.",
                "operationId": "f22e4d82d875d0f5d91ac79b95f7f9e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idPermisos"
                                ],
                                "properties": {
                                    "idPermisos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso(s) aceptado(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "permisoAceptado": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Faltan parámetros requeridos"
                    }
                }
            }
        },
        "/permisos/rechazar": {
            "patch": {
                "tags": [
                    "Permiso"
                ],
                "summary": "Rechazar permisos",
                "description": "Rechaza uno o varios permisos por sus IDs.",
                "operationId": "db22e91f2406a87d1b3db563abac0191",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idPermisos"
                                ],
                                "properties": {
                                    "idPermisos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso(s) rechazado(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "permisoRechazado": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Faltan parámetros requeridos"
                    }
                }
            }
        },
...
    },
...
}